pyproject.toml
[tool.poetry]
name = "cork"
version = "1.0.0"
description = "Another yet ticket board but serious ;)"
authors = ["https://github.com/alx-me-git/cork"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.3"
celery = {extras = ["redis"], version = "^5.3.4"}
psycopg2-binary = "^2.9.8"
python-dotenv = "^1.0.0"
async-timeout = "^4.0.3"
hiredis = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
